openvino/requirements.txt
fastapi==0.103.0
numpy==1.24.4
opencv_contrib_python==4.6.0.66
opencv_python_headless==4.6.0.66
python-dotenv==1.0.0
uvicorn==0.23.2
python-dotenv==1.0.0
python-multipart==0.0.6
rapidocr_openvino==1.3.26
onnxruntime-openvino==1.19.0

